Streaming monix part
Pending: refactor unary calls into a common part
