+++ modules/examples/health-check-fs2/health-client/src/main/resources/application.conf
