+++ modules/examples/health-check-fs2/health-server/src/main/resources/application.conf
